feat(#52): separate exception handling to a new vertical
