+++ android/src/main/java/io/mosip/tuvali/ble/central/ICentralListener.kt
@@ -2,7 +2,7 @@ package io.mosip.tuvali.ble.central
 
 import android.bluetooth.BluetoothDevice
 import android.bluetooth.le.ScanRecord
-import io.mosip.tuvali.openid4vpble.exception.BLEException
+import io.mosip.tuvali.exception.BLEException
 import java.util.*
 
 interface ICentralListener {
