--- android/src/main/java/io/mosip/tuvali/ble/central/exception/CentralStateHandlerException.kt
@@ -1,8 +1,0 @@
-package io.mosip.tuvali.ble.central.exception
-
-import io.mosip.tuvali.openid4vpble.exception.BLEException
-import io.mosip.tuvali.openid4vpble.exception.ErrorCode
-
-class CentralStateHandlerException(message: String, cause: Exception): BLEException(message, cause,
-    ErrorCode.CentralStateHandlerException
-)
