+++ android/src/main/java/io/mosip/tuvali/ble/peripheral/IPeripheralListener.kt
@@ -1,6 +1,6 @@
 package io.mosip.tuvali.ble.peripheral
 
-import io.mosip.tuvali.openid4vpble.exception.BLEException
+import io.mosip.tuvali.exception.BLEException
 import java.util.UUID
 
 interface IPeripheralListener {
