--- android/src/main/java/io/mosip/tuvali/ble/peripheral/exception/PeripheralStateHandlerException.kt
@@ -1,8 +1,0 @@
-package io.mosip.tuvali.ble.peripheral.exception
-
-import io.mosip.tuvali.openid4vpble.exception.BLEException
-import io.mosip.tuvali.openid4vpble.exception.ErrorCode
-
-class PeripheralStateHandlerException(message: String, cause: Exception): BLEException(message, cause,
-    ErrorCode.PeripheralStateHandlerException
-)
