+++ android/src/main/java/io/mosip/tuvali/exception/BLEException.kt
@@ -1,4 +1,4 @@
-package io.mosip.tuvali.openid4vpble.exception
+package io.mosip.tuvali.exception
 
 open class BLEException(message: String, cause: Exception?, val errorCode: ErrorCode): RuntimeException(message, cause) {
 }
