+++ android/src/main/java/io/mosip/tuvali/exception/ExceptionUtils.kt
@@ -1,4 +1,4 @@
-package io.mosip.tuvali.openid4vpble.exception
+package io.mosip.tuvali.exception
 
 class ExceptionUtils {
   companion object {
