+++ android/src/main/java/io/mosip/tuvali/exception/handlers/OpenIdBLEExceptionHandler.kt
@@ -1,13 +1,15 @@
-package io.mosip.tuvali.openid4vpble.exception
+package io.mosip.tuvali.exception.handlers
 
 import android.util.Log
 import com.facebook.react.bridge.Callback
+import io.mosip.tuvali.exception.BLEException
+import io.mosip.tuvali.exception.ErrorCode
+import io.mosip.tuvali.exception.ExceptionUtils
+import io.mosip.tuvali.exception.UnknownException
 
 import io.mosip.tuvali.transfer.Util
-import io.mosip.tuvali.verifier.exception.VerifierException
-import io.mosip.tuvali.verifier.exception.VerifierExceptionHandler
-import io.mosip.tuvali.wallet.exception.WalletException
-import io.mosip.tuvali.wallet.exception.WalletExceptionHandler
+import io.mosip.tuvali.exception.verifier.VerifierException
+import io.mosip.tuvali.exception.wallet.WalletException
 
 class OpenIdBLEExceptionHandler(private val sendError: (String, ErrorCode) -> Unit, private val stopBle: (Callback) -> Unit) {
   private val logTag = Util.getLogTag(javaClass.simpleName)
