+++ android/src/main/java/io/mosip/tuvali/exception/handlers/WalletExceptionHandler.kt
@@ -1,9 +1,10 @@
-package io.mosip.tuvali.wallet.exception
+package io.mosip.tuvali.exception.handlers
 
 import android.util.Log
-import io.mosip.tuvali.openid4vpble.exception.ErrorCode
-import io.mosip.tuvali.openid4vpble.exception.ExceptionUtils
+import io.mosip.tuvali.exception.ErrorCode
+import io.mosip.tuvali.exception.ExceptionUtils
 import io.mosip.tuvali.transfer.Util
+import io.mosip.tuvali.exception.wallet.WalletException
 
 class WalletExceptionHandler(val sendError: (String, ErrorCode) -> Unit) {
   private val logTag = Util.getLogTag(javaClass.simpleName)
