+++ android/src/main/java/io/mosip/tuvali/exception/verifier/TooManyFailureChunksException.kt
@@ -1,0 +1,6 @@
+package io.mosip.tuvali.exception.verifier
+
+import io.mosip.tuvali.exception.BLEException
+import io.mosip.tuvali.exception.ErrorCode
+
+class TooManyFailureChunksException(s: String) : BLEException(s, null, ErrorCode.TooManyFailureChunksException)
