+++ android/src/main/java/io/mosip/tuvali/exception/verifier/VerifierException.kt
@@ -1,0 +1,6 @@
+package io.mosip.tuvali.exception.verifier
+
+import io.mosip.tuvali.exception.BLEException
+import io.mosip.tuvali.exception.ErrorCode
+
+open class VerifierException(message: String,  cause: Exception? = null): BLEException(message, cause, errorCode = ErrorCode.VerifierUnknownException)
