+++ android/src/main/java/io/mosip/tuvali/exception/wallet/MTUNegotiationFailedException.kt
@@ -1,0 +1,6 @@
+package io.mosip.tuvali.exception.wallet
+
+import io.mosip.tuvali.exception.BLEException
+import io.mosip.tuvali.exception.ErrorCode
+
+class MTUNegotiationFailedException(s: String) : BLEException(s, null, ErrorCode.MTUNegotiationException)
