+++ android/src/main/java/io/mosip/tuvali/exception/wallet/WalletException.kt
@@ -1,0 +1,6 @@
+package io.mosip.tuvali.exception.wallet
+
+import io.mosip.tuvali.exception.BLEException
+import io.mosip.tuvali.exception.ErrorCode
+
+open class WalletException(message: String, cause: Exception?): BLEException(message, cause, errorCode = ErrorCode.WalletUnknownException)
