--- android/src/main/java/io/mosip/tuvali/verifier/exception/TooManyFailureChunksException.kt
@@ -1,6 +1,0 @@
-package io.mosip.tuvali.verifier.exception
-
-import io.mosip.tuvali.openid4vpble.exception.BLEException
-import io.mosip.tuvali.openid4vpble.exception.ErrorCode
-
-class TooManyFailureChunksException(s: String) : BLEException(s, null, ErrorCode.TooManyFailureChunksException)
