--- android/src/main/java/io/mosip/tuvali/verifier/exception/VerifierException.kt
@@ -1,6 +1,0 @@
-package io.mosip.tuvali.verifier.exception
-
-import io.mosip.tuvali.openid4vpble.exception.BLEException
-import io.mosip.tuvali.openid4vpble.exception.ErrorCode
-
-open class VerifierException(message: String,  cause: Exception? = null): BLEException(message, cause, errorCode = ErrorCode.VerifierUnknownException)
