--- android/src/main/java/io/mosip/tuvali/verifier/exception/VerifierTransferHandlerException.kt
@@ -1,6 +1,0 @@
-package io.mosip.tuvali.verifier.exception
-
-import io.mosip.tuvali.openid4vpble.exception.BLEException
-import io.mosip.tuvali.openid4vpble.exception.ErrorCode
-
-class VerifierTransferHandlerException(message: String, cause: Exception): BLEException(message, cause, ErrorCode.VerifierTransferHandlerException)
