--- android/src/main/java/io/mosip/tuvali/wallet/exception/MTUNegotiationFailedException.kt
@@ -1,6 +1,0 @@
-package io.mosip.tuvali.wallet.exception
-
-import io.mosip.tuvali.openid4vpble.exception.BLEException
-import io.mosip.tuvali.openid4vpble.exception.ErrorCode
-
-class MTUNegotiationFailedException(s: String) : BLEException(s, null, ErrorCode.MTUNegotiationException)
