--- android/src/main/java/io/mosip/tuvali/wallet/exception/TransferFailedException.kt
@@ -1,6 +1,0 @@
-package io.mosip.tuvali.wallet.exception
-
-import io.mosip.tuvali.openid4vpble.exception.BLEException
-import io.mosip.tuvali.openid4vpble.exception.ErrorCode
-
-class TransferFailedException(s: String) : BLEException(s, null, ErrorCode.TransferFailedException)
