--- android/src/main/java/io/mosip/tuvali/wallet/exception/WalletException.kt
@@ -1,6 +1,0 @@
-package io.mosip.tuvali.wallet.exception
-
-import io.mosip.tuvali.openid4vpble.exception.BLEException
-import io.mosip.tuvali.openid4vpble.exception.ErrorCode
-
-open class WalletException(message: String, cause: Exception?): BLEException(message, cause, errorCode = ErrorCode.WalletUnknownException)
