+++ android/src/main/java/io/mosip/tuvali/wallet/transfer/ITransferListener.kt
@@ -1,6 +1,6 @@
 package io.mosip.tuvali.wallet.transfer
 
-import io.mosip.tuvali.openid4vpble.exception.BLEException
+import io.mosip.tuvali.exception.BLEException
 
 interface ITransferListener {
   fun onResponseSent()
